+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 12.109.0
 
 * `[support]` Added custom checks support
+* `[knf/united]` Improved handling of environment variables
 * `[support]` Improved application info formatting
 * `[support]` Code refactoring